chore(server): rename windows_profile_common to windows_common
Shorter name, same semantics. Linter also trimmed unused imports.
