+++ server/opensandbox_server/services/k8s/windows_profile.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 from typing import Any, Dict, List, Optional
 
 from opensandbox_server.api.schema import PlatformSpec
-from opensandbox_server.services.windows_profile_common import (
+from opensandbox_server.services.windows_common import (
     inject_windows_resource_limits_env,
     inject_windows_user_ports,
     validate_windows_resource_limits,
